♻️ refactor: Migrate `Sentence` and `AudioQuestion` types to `stepTypes` and update imports accordingly
